refactor(gameplay): migrate to Double for physics

+++ src/gameplay/Difficulty.kt
@@ -3,7 +3,7 @@ package gameplay
 import kotlin.math.max
 
 class Difficulty(var score: Int) {
-    fun updateScore(pos: Float) {
+    fun updateScore(pos: Double) {
         score = max(score, pos.toInt())
     }
 }
